Refactor Codex provider into layers package
- Move Codex runtime and app-server probe logic under `provider/Layers`
- Centralize initialize params and provider snapshot types
- Update tests and imports after the Codex provider split
