+++ apps/server/src/provider/Layers/CodexAdapter.ts
@@ -45,7 +45,7 @@ import {
   type CodexSessionRuntimeError,
   type CodexSessionRuntimeOptions,
   type CodexSessionRuntimeShape,
-} from "../codex/CodexSessionRuntime.ts";
+} from "./CodexSessionRuntime.ts";
 import { type EventNdjsonLogger, makeEventNdjsonLogger } from "./EventNdjsonLogger.ts";
 
 const PROVIDER = "codex" as const;
