+++ apps/server/src/provider/Layers/CodexSessionRuntime.ts
@@ -25,11 +25,11 @@ import * as CodexErrors from "effect-codex-app-server/errors";
 import * as CodexRpc from "effect-codex-app-server/rpc";
 import * as EffectCodexSchema from "effect-codex-app-server/schema";
 
-import { buildCodexInitializeParams } from "../codexAppServer.ts";
+import { buildCodexInitializeParams } from "./CodexProvider.ts";
 import {
   CODEX_DEFAULT_MODE_DEVELOPER_INSTRUCTIONS,
   CODEX_PLAN_MODE_DEVELOPER_INSTRUCTIONS,
-} from "./CodexDeveloperInstructions.ts";
+} from "../CodexDeveloperInstructions.ts";
 
 const PROVIDER = "codex" as const;
